Refactor marketplace feature: Remove old marketplace component, add new layout and gem detail screens
- Deleted the old `marketplace.tsx` file and its associated logic.
- Created a new layout file `market_layout.tsx` to manage the marketplace navigation.
- Implemented `gemdetail.tsx` for displaying detailed information about selected gems.
- Updated `index.tsx` to include the new marketplace functionality with search and filtering options.
